backend/RAG/requirements.txt
PyMuPDF
sentence-transformers
numpy
requests
python-dotenv
fastapi
uvicorn
python-multipart
pinecone